minor fixes based on the feedback
