+++ apps/mobile/src/components/UserFollowList/UserFollowCard.tsx
@@ -67,8 +67,6 @@ export function UserFollowCard({ userRef, queryRef, onPress }: UserFollowCardPro
           </View>
         </View>
       </GalleryTouchableOpacity>
-
-      <FollowButton queryRef={query} userRef={user} />
     </View>
   );
 }
